PDA-49 Separating processService into processFormService
