+++ src/main/java/org/entando/plugins/pda/core/service/process/ProcessFormService.java
@@ -1,0 +1,9 @@
+package org.entando.plugins.pda.core.service.process;
+
+import java.util.List;
+import org.entando.plugins.pda.core.engine.Connection;
+import org.entando.plugins.pda.core.model.form.Form;
+
+public interface ProcessFormService {
+    List<Form> getProcessForm(Connection connection, String processId);
+}
